chore(v4): one test file, and trim the docstrings
Folds the kv_indices length-invariance tests into tests/test_dspark.py --
47 tests, one file. Cuts the docstrings on the three attention halves and
their ops roughly in half; the reasoning that earned its place stays, the
essays do not.
